Refactor InventoryDataProvider and make all the related apis experimental
